✨ feat [#55] : DeliveryManager 반환 로직 구현
- Feign Client 설정 추가
- sequence 타입 Integer -> Long 변경
- 배송담당자 배정 엔드포인트 구현
- 배송담당자 배정 비지니스 로직 구현
- 출발HubId와 타입으로 적절한 배송담당자 조회하는 쿼리
- 최대 순번 조회하는 쿼리

+++ com.sparta.logistics.delivery-manager-service/src/main/java/com/sparta/logistics/delivery_service/application/mapper/DeliveryManagerMapper.java
@@ -6,7 +6,7 @@ import com.sparta.logistics.delivery_service.domain.model.DeliveryManager;
 
 public class DeliveryManagerMapper {
 
-    public static DeliveryManager toEntity(DeliveryManagerCreateRequestDto dto, Long id, Integer sequence) {
+    public static DeliveryManager toEntity(DeliveryManagerCreateRequestDto dto, Long id, Long sequence) {
         return DeliveryManager.builder()
                 .id(id)
                 .hubId(dto.getHubId())
